pyproject.toml
[project]
name = "ai-model-quality-framework"
version = "0.1.0"
description = "Production-style ML system under test: Fake News Detection + AI testing, validation, drift monitoring, and CI/CD gates."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.4",
  "scipy>=1.12",
  "joblib>=1.3",
  "pyyaml>=6.0",
  "pydantic>=2.6",
  "fastapi>=0.110",
  "uvicorn>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.1",
  "ruff>=0.6",
  "mypy>=1.8",
  "pre-commit>=3.6",
  "rich>=13.7",
  "pyarrow>=15.0",

]
nlp = [
  "nltk>=3.8",
  "regex>=2024.0.0",
]
monitoring = [
  "evidently>=0.4.0",
]
hf = [
  "transformers>=4.41",
  "datasets>=2.19",
  "evaluate>=0.4.2",
  "accelerate>=0.31",
  "torch>=2.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
